requirements.txt
Django==2.0
gunicorn==19.7.1
djangorestframework==3.7.7

# Below celery is the result of celery[redis]
celery==4.1.0

# Development
ipython
